+++ java/src/main/kotlin/io/spine/validation/java/option/PatternFieldGenerator.kt
@@ -47,13 +47,13 @@ import io.spine.validation.PatternField
 import io.spine.validation.isRepeatedString
 import io.spine.validation.isSingularString
 import io.spine.validation.java.expression.ConstraintViolationClass
-import io.spine.validation.java.ErrorPlaceholder
-import io.spine.validation.java.ErrorPlaceholder.FIELD_PATH
-import io.spine.validation.java.ErrorPlaceholder.FIELD_TYPE
-import io.spine.validation.java.ErrorPlaceholder.FIELD_VALUE
-import io.spine.validation.java.ErrorPlaceholder.PARENT_TYPE
-import io.spine.validation.java.ErrorPlaceholder.REGEX_MODIFIERS
-import io.spine.validation.java.ErrorPlaceholder.REGEX_PATTERN
+import io.spine.validation.java.violation.ErrorPlaceholder
+import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_PATH
+import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_TYPE
+import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_VALUE
+import io.spine.validation.java.violation.ErrorPlaceholder.PARENT_TYPE
+import io.spine.validation.java.violation.ErrorPlaceholder.REGEX_MODIFIERS
+import io.spine.validation.java.violation.ErrorPlaceholder.REGEX_PATTERN
 import io.spine.validation.java.FieldOptionCode
 import io.spine.validation.java.expression.FieldPathClass
 import io.spine.validation.java.expression.ImmutableListClass
@@ -62,11 +62,11 @@ import io.spine.validation.java.expression.StringClass
 import io.spine.validation.java.ValidationCodeInjector.MessageScope.message
 import io.spine.validation.java.ValidationCodeInjector.ValidateScope.parentPath
 import io.spine.validation.java.ValidationCodeInjector.ValidateScope.violations
-import io.spine.validation.java.expression.constraintViolation
-import io.spine.validation.java.fieldPath
-import io.spine.validation.java.joinToString
+import io.spine.validation.java.violation.constraintViolation
+import io.spine.validation.java.violation.fieldPath
+import io.spine.validation.java.violation.joinToString
 import io.spine.validation.java.mangled
-import io.spine.validation.java.expression.templateString
+import io.spine.validation.java.violation.templateString
 import java.util.regex.Matcher
 import java.util.regex.Pattern
 import org.apache.commons.lang.StringEscapeUtils.escapeJava
