+++ java/src/main/kotlin/io/spine/validation/java/option/RequiredFieldGenerator.kt
@@ -38,17 +38,17 @@ import io.spine.validate.ConstraintViolation
 import io.spine.validation.IF_MISSING
 import io.spine.validation.RequiredField
 import io.spine.validation.java.expression.DefaultValueExpression
-import io.spine.validation.java.ErrorPlaceholder
-import io.spine.validation.java.ErrorPlaceholder.FIELD_PATH
-import io.spine.validation.java.ErrorPlaceholder.FIELD_TYPE
-import io.spine.validation.java.ErrorPlaceholder.PARENT_TYPE
+import io.spine.validation.java.violation.ErrorPlaceholder
+import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_PATH
+import io.spine.validation.java.violation.ErrorPlaceholder.FIELD_TYPE
+import io.spine.validation.java.violation.ErrorPlaceholder.PARENT_TYPE
 import io.spine.validation.java.FieldOptionCode
 import io.spine.validation.java.ValidationCodeInjector.ValidateScope.parentPath
 import io.spine.validation.java.ValidationCodeInjector.ValidateScope.violations
-import io.spine.validation.java.expression.constraintViolation
-import io.spine.validation.java.fieldPath
-import io.spine.validation.java.joinToString
-import io.spine.validation.java.expression.templateString
+import io.spine.validation.java.violation.constraintViolation
+import io.spine.validation.java.violation.fieldPath
+import io.spine.validation.java.violation.joinToString
+import io.spine.validation.java.violation.templateString
 
 /**
  * The generator for `(required)` option.
