+++ java/src/main/kotlin/io/spine/validation/java/violation/FieldPathExpressions.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.java
+package io.spine.validation.java.violation
 
 import io.spine.base.FieldPath
 import io.spine.protodata.ast.FieldName
